Revert "Merge pull request NixOS/nixpkgs#73251 which removes SLiM DM"
This reverts commit 4583e293fd5f4df1246b658d802a1b519bc05cd8,
reversing changes made to d6fa5402bb7d35a7e7d0574d61a271c7e88ce079.

The fact that SLiM is not officially actively maintained is not a good
reason to remove it. It works, and, in some instances, like automatic
logins, it is the only thing that actually works. See
https://github.com/NixOS/nixpkgs/pull/46396 and
https://github.com/NixOS/nixpkgs/pull/30890#issuecomment-419670755.