linux-libre: set update-libre.sh as update script
This doesn't actually update the kernel, just the linux-libre
deblobbing scripts, but it should mean that automatic updaters keep
the deblobbing scripts up to date.  So even if deblobbing scripts for
a new kernel version are not available immediately after release, they
should be updated automatically soon enough once available.