nixos/iwd: add `networks` and `interfaces` option
This change adds support for declarative networks with `iwd(8)`, similar
to `networking.wireless.networks` and an `interfaces` option which tells
the service which to use (by default, the interfaces will be found
automatically).

In contrast to the `wpa_supplicant` module it's still possible to
add additional networks imperatively using `iwctl(1)` since all networks
are stored as separated files in `/var/lib/iwd`.

Please note that this change doesn't add support for 802.1x networks[1]
as this would currently mean that we'd have to store sensitive secrets
for those networks world-readably in the Nix-store. It's recommended to
add those networks imperatively.

This module supports open networks and networks with a WPA PSK or a raw
passphrase. If a SSID doesn't contain alphanumeric chars and `-_` only,
the network-configuration will be stored as `=<hex-encoded-version>.<ext>`
in `/var/lib/iwd`.

[1] https://iwd.wiki.kernel.org/networkconfigurationsettings#x_wpawpa2_enterprise_settings