qt5.qtbase: python2 -> python3
The only python reference in qtbase's dev output is a python script
which imports print from __future__ (mkspecs/features/uikit/devices.py).
It's not clear whether it depends on python or it just provides a python
script. But I believe it's still good to have in buildInputs so it has
a proper shebang.

Co-authored-by: worldofpeace <worldofpeace@protonmail.ch>