glirc: init at 2.33.1
haskellPackages.glirc is a "Console IRC Client." I've added a doJailbreak
to fix the build (thanks @infinisil) and added it to top-level/all-packages.nix
so people can find and install it as they would normally.

Would be nice to make this build in a way that allows the OTR extension to be
enabled. One thing at a time....