python: change alias from python2 to python3
The end-of-life for python2 is coming up fast (1st of January 2020) and
as such I think we should be trying as hard as we can to discourage its
use in nixpkgs. Hopefully, anything using the unversioned alias
shouldn't depend too much on which version of python it winds up with,
but I'm sure this is going to break an awful lot of packages.