pantheon.wingpanel-with-indicators: have indicators argument only append to defaults
Wingpanel was designed firstly as an indicator renderer,
and as such just a container for indicators that are distributed
outside itself. Being able to control which and each indicator with
`indicators` is confusing, ideally each of the default indicators
would be shipped with wingpanel itself. I don't see how this
kind of extensibility would be useful to a user so we're going
to append to the expected defaults. The `useDefaultIndicators`
argument is there to development test a single indicator/s.