bat: insert less into PATH
The default pager for `bat` is `less -RF --no-init`. Overriding this to
`less -RF` in order to enable alternate screen behavior (e.g. page
up/down keys) causes the `-F` flag to behave badly in the version of
`less` shipped on macOS. The simplest way to fix this is to simply put
Nix's `less` into the `PATH` so it takes precedence.