root: wrap python executables
In normal use of ROOT the PYTHONPATH is intended to be set when user
sources the thisroot.{,s}sh. We do that in the setupHook. This covers
the case when thisroot.sh was not sourced.

(cherry picked from commit 0b28198f23fab730a797e865e741d9fd8a50f1a3)

cc #76541