haskell-ghcide: drop broken overrides
We made an effort to support ghcide in Nixpkgs, but the complexity of the
problem is a bit too high, IMHO. We need to keep older versions of several
packages around in order to satisfy the build requirements, and some of those
older packages don't even build themselves (like hie-bios). We had ghcide
working at some point, but then it was broken again right away after a couple
of days. I fear that we'll run into that issue again and again with a setup of
that complexity.

Instead, I'd propose that we work with upstream to fix their build, i.e. let's
make sure that the proper ghcide build works with recent versions of its build
inputs.

Closes https://github.com/NixOS/nixpkgs/pull/75449.
Closes https://github.com/NixOS/nixpkgs/pull/76103.