nixos/transmission: fix from where service loads config
Currently if you specify home to be someplace else than ~/ for user
then Transmissions always attempts to load the config from the
default location which is $HOME/.config/transmission-daemon based on documentation:
https://github.com/transmission/transmission/wiki/Configuration-Files

Which means that the changes done to the config under settingsDir in
ExecPreStart have no effect because they are modifying a file that is never loaded.

I've added an explicit --config-dir ${settingsDir} to make sure
that Transmission loads the correct config file even when home is changed.

Signed-off-by: Jakub Sokołowski <jakub@status.im>