stdenv.mkDerivation: add support for env attrset
For `__structuredAttrs`, we want to distinguish regular structured variables
from string-coercible variables that will be exported to the environment.
The environment variables will be passed to `mkDerivation` as attributes
of the `env` attribute set.

To lower the review workload of the `__structuredAttrs` PR, we have decided
to make the `env` mechanism available as a vanguard. Until the support for
`__structuredAttrs` is implemented, the content of the `env` attrset will
be merged into the attrset passed to `mkDerivation`.

The content of the attributes in `env` is checked for validity and conflicts
with the rest of `mkDerivation` attributes. The current mechanism should
produce the same derivations as the original, thus require no rebuilds.

This change is backwards compatible for now, as Nix does not allow attrsets
to be passed to derivation without `__structuredAttrs` enabled, and the current
mechanism handles the non-attrsets as before. We reserve the right to require
`env` to be attrset in the future.