flyway: fix jarDirs option
this is not a system property, but rather an argument to
org.flaywaydb.commandline.Main.main().  therefore, it must come after,
rather than before, the name of the main class.  (otherwise it's
interpreted as an argument to the VM itself.)

flyway also expects the option and value to be separated by '=', not ' '.

follows on to #76094