taffybar: fix gi-gdkx11 dependency
haskellPackages chooses to default to the newer (4.X) gtk4 versions,
even though nix doesn't provide the necessary libraries.

This change makes sure that the gtk3 version of gi-gdkx11 will be
available after the next package update, and provides an override so
that taffybar will use it and thus be compilable.