release: drop nox for nixpkgs-review
The comment in this file about why nox is here
was "Needed by travis-ci to test PRs", and we
for sure don't use travis-ci anymore.
Instead we're making nixpkgs-review a deliverable
because it's included in the PR template as a tool
to use. We also swapped out nox for nixpkgs-review
in the PR template, so I believe this makes sense
here because it was changed.