gnustep/base: add libiberty to linker flags
In https://github.com/NixOS/nixpkgs/issues/76927, we found that
xmemdup is missing when compiling some conftest for gnustep-base. It’s
unclear why this happens but it appears to be from the update to gcc9.
Adding libiberty gives us that symbol and fixes things. What’s less
clear is why this happens to begin with.

Fixes #76927