clang: fix compilation of HIP-code
When compiling HIP-code clang executes other llvm binaries, such as
llvm-link, llc or opt. Clang expects these binaries in the folder with
the clang-executable (...clang/bin/), this is not true for nix, because
these binaris are in the llvm/bin folder. It is not possible to simply
replace the path of the called library, because then the compilation
with gcc fails. While I personally did not encounter a problem with
clang failing due to calling lld I added it for completeness, this
however means clang also depends on the llvm binutils.