blitz: remove package
This is very out of date, is pinned to an old version of boost, and the last
update was 2.5 years ago marking it as broken. Since no one has fixed it or
upgraded to the least version, it seems a strong indication it's no longer
used/wanted/maintained and should be removed.