hfsprogs: 332.25 -> 540.1.linux3
This is what most distributions ship for `hfsprogs`,
significantly newer than 332.25 (10.4.11[1], 2005) but only
about half as old (first appears in 10.7[2], 2011).

The newer tools drop support for `mkfs.hfs` (creating an HFS partition)
however they retain the ability to check both HFS and HFS+.

This seems like a small cost, and can be addressed by either reviving
the 332.25 variant or using hfsutils[3][4].  Coming soon, maybe? :D

It seems HFS drives and images are already unsupported and have been
for quite some time, so while it is unfortunate to drop support
the filesystem creators Apple apparently are very done with it:

"With the introduction of Mac OS X 10.6, Apple dropped support for
formatting or writing HFS disks and images, which remain supported as
read-only volumes.  Starting with macOS 10.15, HFS disks can no longer
be read. "[5]

[1] https://opensource.apple.com/release/mac-os-x-10411x86.html
[2] https://opensource.apple.com/release/mac-os-x-107.html
[3] https://www.mars.org/home/rob/proj/hfs/
[4] https://packages.debian.org/source/sid/hfsutils (for patches)
[5] https://en.wikipedia.org/wiki/Hierarchical_File_System