firebird: mark as broken
Hasn't been updated since 2017 and breaks when building with a more
recent gcc (gcc5 needs to be removed as it breaks with glibc 2.30). I
tried to update this to a more recent version which didn't work either,
but as this is pretty old I'm not even sure if this package is used atm.