r-packages: update
This update was primarily done to update rPackages.V8 to 3.0 which
doesn't depend on an ancient version of v8 anymore.

Also dropped the `-lv8_libplatform` linker flag. It seems as this now
part of `v8.so` as `v8_libplatform.so` doesn't exist anymore on recent
v8 versions in nixpkgs, but the headers are still there and there aren't
any "undefined reference to" errors while linking.