ums: 6.2.2 -> 9.1.0
Removing the packaged JRE so that `jre8` gets used.

Patchelf fails on `tsMuxeR` and `tsMuxeR-new`, but this doesn't make
it any worse comparing to 6.2.2 where is also failed, so I am assuming
fixing it is out of scope of this pr. The only difference is that the
new UMS version correctly identifies that `tsMuxeR` doesn't work,
comparing to the previous one.