clang_7,clang_8,clang_9: fix compilation of HIP-code
This PR aims to fix the same problem as NixOS/nixpkgs PR #77476; enabling to
compile HIP-code using the packaged clang compiler, by also searching in $PATH
for required binaries.
The change is committed upstream (https://reviews.llvm.org/D72903), but will
not land in the clang versions in nixpkgs (only clang 10+). As such I have
created patches for the affected versions. To compile HIP-code lld is needed,
so I added it to the clang-package.<Paste>