buildRustCrate: dont clobber src /w gen code
If a crate happened to have a build script that generates a file with
the name that’s the same with another file in `src/`, this little
snippet of code would overwrite the source with generated code…

Having looked at the history I failed to find justification for this
behaviour and since it makes actual crates (rustversion) fail to build,
this snippet has to go. See ya!