python3Packages.python-ly: remove myself from maintainer list
While going through my repology feed of package-updates I realized that
I'm still a maintainer of this. Just did an update in e6a9c15fa1d86045549a6e385e2621945419850c
and the package is pretty easy to maintain, but decided to make it clear
that I don't actively maintain this.