wayland: 1.17.0 -> 1.18.0
This also switches the build from Autotools to Meson. The only
difference should be that the libtool archives (.la files) aren't
generated anymore.
Additionally I've updated the meta attributes and added myself as
maintainer.

The changelog is in the final release announcement [0] and an overview
of most commits is in the alpha release announcement [1].

[0]: https://lists.freedesktop.org/archives/wayland-devel/2020-February/041207.html
[1]: https://lists.freedesktop.org/archives/wayland-devel/2020-January/041140.html