pythonPackages.pyopengl: fix runtime shared library loading failure
This closes #76822.

pyopengl 3.1.4 introduced a new logic for shared library loading: it
tests a few combinations of library name and suffix (such as .so.X).

Our previous patch was just replacing the library name (e.g. 'glut') by
the full path to the nix store. This does not work anymore with pyopengl
3.1.4 new heuristic.

This commit just keep the behavior of pyopengl but adds the nix store
path to the list of tried paths.