fix canon-cups-ufr2 for previously broken printers
The ufr2 driver relies on black-box binaries from Canon, with baked-in
FHS assumptions. Previously, a `LD_PRELOAD` shim has been used to adapt
calls to some libc functions and correct the path arguments passed.
Unfortunately, this hack doesn't work for all printers; some of the
Canon libraries rely on invoking syscalls directly, and can't be
mollified with `ld.so` tricks.

And so blacker arts were needed. Introduce a wrapper that uses `ptrace`
to intercept all syscalls performed by the binary; when a supported
syscall is identified, read its path argument from the binary's address
space, see if it's a broken FHS path, and if necessary inject a
corrected path into the process's address space before allowing the
fixed-up syscall to proceed.

This has some pros and cons. The cons:
- `ptrace` is sometimes restricted (it's allowed under the default NixOS
  config though)
- there's a minor performance penalty (syscalls cost several times as
  much)
- it's crazy (this binary doesn't seem to do any of the things that
  might break it, though)
The pros:
- got my D530 working (just in time to print that shipping label...)
- it's a general solution; it should work for any printer that uses this
  driver, and it should work for future versions of the driver